predictor/requirements.txt
backports.statistics==0.1.0
fastapi==0.116.1
numpy==2.3.3
pandas==2.3.2
prophet==1.1.7
pytz==2025.1
pytz==2025.2
supabase==2.18.1
uvicorn==0.35.0
